Rename table client headers and split out the clients into their own files.
