+++ sdk/tables/azure-data-tables/CMakeLists.txt
@@ -49,7 +49,8 @@ set(
     inc/azure/data/tables/models.hpp
     inc/azure/data/tables/rtti.hpp
     inc/azure/data/tables.hpp
-    inc/azure/data/tables/tables_clients.hpp
+    inc/azure/data/tables/table_client.hpp
+    inc/azure/data/tables/table_service_client.hpp
     inc/azure/data/tables/tables_sas_builder.hpp
 )
