+++ sdk/tables/azure-data-tables/test/ut/serializers_test.hpp
@@ -3,11 +3,10 @@
 // Licensed under the MIT License.
 
 #include "../src/private/serializers.hpp"
-#include "azure/data/tables/tables_clients.hpp"
+#include "azure/data/tables/table_client.hpp"
 #include "test/ut/test_base.hpp"
 
 namespace Azure { namespace Data { namespace Test {
 
-  class SerializersTest : public Azure::Storage::Test::StorageTest {
-  };
+  class SerializersTest : public Azure::Storage::Test::StorageTest {};
 }}} // namespace Azure::Data::Test
